+++ impl/src/planning/subplanners.rs
@@ -35,12 +35,7 @@ use crate::{
   metadata::RazeMetadata,
   planning::license,
   settings::{format_registry_url, CrateSettings, GenMode, RazeSettings},
-  util::get_package_ident,
-  util::get_workspace_member_path,
-  util::{
-    self, filter_bazel_triples, generate_bazel_conditions, get_matching_bazel_triples,
-    is_bazel_supported_platform, PlatformDetails,
-  },
+  util,
 };
 
 use super::{
@@ -86,7 +81,7 @@ struct TargetedDependencySet {
 struct CrateSubplanner<'planner> {
   // Workspace-Wide details
   settings: &'planner RazeSettings,
-  platform_details: &'planner Option<PlatformDetails>,
+  platform_details: &'planner Option<util::PlatformDetails>,
   crate_catalog: &'planner CrateCatalog,
   // Crate specific content
   crate_catalog_entry: &'planner CrateCatalogEntry,
@@ -99,7 +94,7 @@ struct CrateSubplanner<'planner> {
 /** An internal working planner for generating context for a whole workspace. */
 pub struct WorkspaceSubplanner<'planner> {
   pub(super) settings: &'planner RazeSettings,
-  pub(super) platform_details: &'planner Option<PlatformDetails>,
+  pub(super) platform_details: &'planner Option<util::PlatformDetails>,
   pub(super) crate_catalog: &'planner CrateCatalog,
   pub(super) metadata: &'planner RazeMetadata,
 }
@@ -137,7 +132,10 @@ impl<'planner> WorkspaceSubplanner<'planner> {
           if self.settings.binary_deps.contains_key(&pkg.name) {
             None
           } else {
-            get_workspace_member_path(&pkg.manifest_path, &self.metadata.metadata.workspace_root)
+            util::get_workspace_member_path(
+              &pkg.manifest_path,
+              &self.metadata.metadata.workspace_root,
+            )
           }
         } else {
           None
@@ -279,8 +277,8 @@ impl<'planner> CrateSubplanner<'planner> {
     // Build a list of dependencies while addression a potential whitelist of target triples
     let mut filtered_deps = Vec::new();
     for dep_set in targeted_deps.iter() {
-      let mut target_triples = get_matching_bazel_triples(&dep_set.target)?;
-      filter_bazel_triples(
+      let mut target_triples = util::get_matching_bazel_triples(&dep_set.target)?;
+      util::filter_bazel_triples(
         &mut target_triples,
         self
           .settings
@@ -303,7 +301,7 @@ impl<'planner> CrateSubplanner<'planner> {
           dev_dependencies: dep_set.dependencies.dev_deps.clone(),
           aliased_dependencies: dep_set.dependencies.aliased_deps.clone(),
         },
-        conditions: generate_bazel_conditions(
+        conditions: util::generate_bazel_conditions(
           &self.settings.rust_rules_workspace_name,
           &target_triples,
         )?,
@@ -324,7 +322,7 @@ impl<'planner> CrateSubplanner<'planner> {
           .iter()
           .find(|pkg| pkg.id == *pkg_id);
         if let Some(package) = workspace_member {
-          get_workspace_member_path(
+          util::get_workspace_member_path(
             &package.manifest_path,
             &self.crate_catalog.metadata.workspace_root,
           )
@@ -416,7 +414,7 @@ impl<'planner> CrateSubplanner<'planner> {
         .package();
 
       // Skip settings-indicated deps to skip
-      if all_skipped_deps.contains(&get_package_ident(
+      if all_skipped_deps.contains(&util::get_package_ident(
         &dep_package.name,
         &dep_package.version.to_string(),
       )) {
@@ -545,7 +543,8 @@ impl<'planner> CrateSubplanner<'planner> {
           }
         }
 
-        let (is_bazel_platform, matches_all_platforms) = is_bazel_supported_platform(&target_str);
+        let (is_bazel_platform, matches_all_platforms) =
+          util::is_bazel_supported_platform(&target_str);
         // If the target is not supported by Bazel, we ignore it
         if !is_bazel_platform {
           continue;
