+++ impl/src/testing.rs
@@ -12,19 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use cargo_metadata::{Metadata, PackageId};
 use flate2::Compression;
 use httpmock::{Method::GET, MockRef, MockServer};
 use indoc::{formatdoc, indoc};
+use semver::Version;
 use serde_json::json;
+use tempfile::TempDir;
+
 use std::{
   collections::HashMap,
   fs::{create_dir_all, write, File},
   io::Write,
   path::Path,
 };
-use tempfile::TempDir;
 
-use crate::{metadata::CargoWorkspaceFiles, util::get_package_ident};
+use crate::{
+  metadata::{tests::dummy_raze_metadata, CargoWorkspaceFiles},
+  util::get_package_ident,
+};
 
 pub const fn basic_toml_contents() -> &'static str {
   indoc! { r#"
@@ -286,4 +292,42 @@ pub fn mock_crate_index(
   } else {
     None
   }
+}
+
+/** Generate some basic metadata with */
+pub fn dummy_modified_metadata() -> Metadata {
+  let mut metadata = dummy_raze_metadata().metadata.clone();
+
+  // Phase 1: Add a dummy dependency to the dependency graph.
+
+  let mut resolve = metadata.resolve.take().unwrap();
+  let mut new_node = resolve.nodes[0].clone();
+  let name = "test_dep";
+  let name_id = "test_dep_id";
+
+  // Add the new dependency.
+  let id = PackageId {
+    repr: name_id.to_string(),
+  };
+  resolve.nodes[0].dependencies.push(id.clone());
+
+  // Add the new node representing the dependency.
+  new_node.id = id;
+  new_node.deps = Vec::new();
+  new_node.dependencies = Vec::new();
+  new_node.features = Vec::new();
+  resolve.nodes.push(new_node);
+  metadata.resolve = Some(resolve);
+
+  // Phase 2: Add the dummy dependency to the package list.
+
+  let mut new_package = metadata.packages[0].clone();
+  new_package.name = name.to_string();
+  new_package.id = PackageId {
+    repr: name_id.to_string(),
+  };
+  new_package.version = Version::new(0, 0, 1);
+  metadata.packages.push(new_package);
+
+  metadata
 }
